+++ CHANGES.rst
@@ -8,6 +8,7 @@ New Features:
 
 * UserStrings: arbitrary user-defined strings that can be defined in the
   template configuration
+* strings in a StringCollection can now be styled text
 * Sphinx frontend: use the `today` and `today_fmt` configuration variables for
   the date on the title page
 * Sphinx frontend: allow extensions access to the builder object (issue #155)